refactor: Move header into its own component
