+++ src/App.js
@@ -1,5 +1,7 @@
-import React from 'react';
+/*global chrome*/
+import React, { useEffect, useState} from 'react';
 import TabList from './components/TabList';
+import Header from './components/Header';
 import './App.css';
 import { makeStyles } from '@material-ui/core/styles';
 
@@ -28,11 +30,34 @@ const useStyles = makeStyles(() => ({
 
 function App() {
   const classes = useStyles();
+  const [tabList, setTabList] = useState([]);
+
+  useEffect(() => {
+    chrome.windows.getAll({populate:true},function(windows){
+      const tabs = []
+      windows.forEach(function(window){
+        window.tabs.forEach(function(tab){
+          if(tab.audible) {
+            tabs.push({
+              tab,
+              windowId: window.id
+            });
+          }
+        });
+      });
+      setTabList(tabs);
+    });
+  })
 
   return (
     <div className="App">
       <header className={classes.root}>
-        <TabList />
+        <Header
+          tabList={tabList}
+        />
+        <TabList 
+          tabList={tabList}
+        />
       </header>
     </div>
   );
